# backend/internal/handler/v1/swagger/swagger.yaml
definitions:
...
  dd-prediction-api_internal_model.CreateCryptoDuelResp:
    properties:
      duel:
        $ref: '#/definitions/dd-prediction-api_internal_model.Duel'
      result:
        $ref: '#/definitions/dd-prediction-api_internal_model.TxHashResp'
    type: object
  dd-prediction-api_internal_model.CreateDuelReq:
    properties:
      answer:
        type: integer
      commission_rate:
        type: integer
      deadline:
        type: string
      duel_info:
        additionalProperties: {}
        type: object
      duel_price:
        type: number
      is_owner_resolving:
        type: boolean
      logo_url:
        type: string
      max_price:
        type: number
      min_price:
        type: number
      price_type:
        type: string
      question:
        type: string
      source_of_truth:
        type: string
      symbol:
        type: string
    type: object
...
  dd-prediction-api_internal_model.JoinCryptoDuelResp:
    properties:
      player:
        $ref: '#/definitions/dd-prediction-api_internal_model.Player'
      result:
        $ref: '#/definitions/dd-prediction-api_internal_model.TxHashResp'
    type: object
  dd-prediction-api_internal_model.JoinDuelReq:
    properties:
      answer:
        type: integer
      duel_id:
        type: string
      external_source:
        type: string
      invited_by:
        type: string
      multi_duel_id:
        type: string
      multi_duel_referral_code:
        type: string
      paid_price:
        type: number
    type: object
...
  dd-prediction-api_internal_model.TxHashResp:
    properties:
      tx_hash:
        type: string
    type: object
...
  dd-prediction-api_pkg_apperrors.ErrorPublic:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a swagger specification for a Duel Duck back-end.
  title: Duel Duck Prediction API
  version: "1.0"
paths:
...
  /admin/duel:
    post:
      consumes:
      - application/json
      description: Creates a new crypto duel as an admin. The duel is immediately
        set to active status, bypassing the approval step. The action is recorded
        in the audit log.
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Duel creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dd-prediction-api_internal_model.CreateDuelReq'
      produces:
      - application/json
      responses:
        "200":
          description: Created duel
          schema:
            $ref: '#/definitions/dd-prediction-api_internal_model.Duel'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "403":
          description: Admin permission required
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
      security:
      - BearerAuth: []
      summary: Create a crypto duel (admin)
      tags:
      - admin
...
  /crypto-duel/external-wallet/solana:
    post:
      consumes:
      - application/json
      description: Creates an application-based crypto duel using the provided data.
        Requires authentication.
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Duel created successfully
          schema:
            $ref: '#/definitions/dd-prediction-api_internal_model.CreateCryptoDuelResp'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "401":
          description: Unauthorized - Invalid or missing claims
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
      security:
      - BearerAuth: []
      summary: Create app-based crypto duel
      tags:
      - duel
  /crypto-duel/external-wallet/solana/join:
    post:
      consumes:
      - application/json
      description: Allows an authenticated user to join an existing app-based crypto
        duel.
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Joined the duel successfully
          schema:
            $ref: '#/definitions/dd-prediction-api_internal_model.JoinCryptoDuelResp'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "401":
          description: Unauthorized - Invalid or missing claims
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
      security:
      - BearerAuth: []
      summary: Join an app-based crypto duel
      tags:
      - duel
  /crypto-duel/external-wallet/solana/join/sign-tx:
    post:
      consumes:
      - application/json
      description: Creates a signed transaction that allows a user to join an app-based
        crypto duel.
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Duel join request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dd-prediction-api_internal_model.JoinDuelReq'
      produces:
      - application/json
      responses:
        "200":
          description: Signed transaction returned successfully
          schema:
            properties:
              tx:
                type: string
            type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "401":
          description: Unauthorized - Invalid or missing claims
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
      security:
      - BearerAuth: []
      summary: Sign transaction to join an app crypto duel
      tags:
      - duel
  /crypto-duel/external-wallet/solana/sign-tx:
    post:
      consumes:
      - application/json
      description: Creates and signs a transaction for initializing an application-based
        crypto duel. Returns a signed transaction that can be submitted to the blockchain.
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request data for creating a duel
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dd-prediction-api_internal_model.CreateDuelReq'
      produces:
      - application/json
      responses:
        "200":
          description: Signed transaction returned successfully
          schema:
            properties:
              tx:
                type: string
            type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "401":
          description: Unauthorized - Invalid or missing claims
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
      security:
      - BearerAuth: []
      summary: Sign transaction to create app crypto duel
      tags:
      - duel
  /crypto-duel/self-resolve:
    post:
      consumes:
      - application/json
      description: Allows the duel owner to resolve an owner-resolving crypto duel.
        Returns resulting transaction hashes.
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Duel resolve request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dd-prediction-api_internal_model.DuelResolveReq'
      produces:
      - application/json
      responses:
        "200":
          description: Transaction hashes and duel
          schema:
            $ref: '#/definitions/dd-prediction-api_internal_model.ResolveCryptoDuelResp'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "401":
          description: Unauthorized - Invalid or missing claims
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
      security:
      - BearerAuth: []
      summary: Resolve an owner-resolving duel  crypto duel
      tags:
      - duel
  /crypto-duel/token-accounts:
    get:
      consumes:
      - application/json
      description: Retrieves the balances of all token accounts associated with the
        user's public blockchain address. Requires authentication. Supports filtering
        and pagination via query parameters.
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - default: '"token"'
        description: Type of token account to filter (e.g. SPL, NFT, etc.)
        in: query
        name: type
        type: string
      - default: 1
        description: Page number for pagination (starts from 1)
        in: query
        name: page_number
        type: integer
      - default: 20
        description: Number of token accounts per page
        in: query
        name: page_size
        type: integer
      - description: Hide token accounts with zero balance
        in: query
        name: hide_zero
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Token account balances retrieved successfully
          schema:
            $ref: '#/definitions/dd-prediction-api_internal_model.SolscanTokenAccountsResp'
        "400":
          description: Invalid request data or query parameters
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "401":
          description: Authentication required or invalid token
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "503":
          description: External service unavailable
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
      security:
      - BearerAuth: []
      summary: Get token account balances for the user's public address
      tags:
      - duel
...
  /duel/public/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a duel by its ID without requiring authentication. Returns
        the duel details if found.
      parameters:
      - description: Duel ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Duel details
          schema:
            $ref: '#/definitions/dd-prediction-api_internal_model.Duel'
        "400":
          description: Invalid duel ID
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "404":
          description: Duel not found
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
      summary: Get duel by ID (public)
      tags:
      - duel
...
securityDefinitions:
  BearerAuth:
    description: Enter the auth token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"